Reference the CommonJS helpers when a file is not explicitly a module.
