+++ packages/babel-plugin-transform-runtime/test/fixtures/use-options/corejs-useES6Modules/input.mjs
