--- packages/babel-plugin-transform-runtime/test/fixtures/use-options/corejs-useES6Modules/output.js
@@ -1,23 +1,0 @@
-var _classCallCheck = require("@babel/runtime-corejs2/helpers/esm/classCallCheck");
-
-var _possibleConstructorReturn = require("@babel/runtime-corejs2/helpers/esm/possibleConstructorReturn");
-
-var _getPrototypeOf = require("@babel/runtime-corejs2/helpers/esm/getPrototypeOf");
-
-var _inherits = require("@babel/runtime-corejs2/helpers/esm/inherits");
-
-let Foo =
-/*#__PURE__*/
-function (_Bar) {
-  "use strict";
-
-  _inherits(Foo, _Bar);
-
-  function Foo() {
-    _classCallCheck(this, Foo);
-
-    return _possibleConstructorReturn(this, _getPrototypeOf(Foo).apply(this, arguments));
-  }
-
-  return Foo;
-}(Bar);
