+++ packages/babel-plugin-transform-runtime/test/fixtures/use-options/corejs/input.mjs
