+++ packages/babel-plugin-transform-runtime/test/fixtures/use-options/corejs/output.mjs
@@ -1,0 +1,18 @@
+import _classCallCheck from "@babel/runtime-corejs2/helpers/classCallCheck";
+import _possibleConstructorReturn from "@babel/runtime-corejs2/helpers/possibleConstructorReturn";
+import _getPrototypeOf from "@babel/runtime-corejs2/helpers/getPrototypeOf";
+import _inherits from "@babel/runtime-corejs2/helpers/inherits";
+
+let Foo =
+/*#__PURE__*/
+function (_Bar) {
+  _inherits(Foo, _Bar);
+
+  function Foo() {
+    _classCallCheck(this, Foo);
+
+    return _possibleConstructorReturn(this, _getPrototypeOf(Foo).apply(this, arguments));
+  }
+
+  return Foo;
+}(Bar);
