+++ packages/babel-plugin-transform-runtime/test/fixtures/use-options/useESModules-cjs/input.js
