+++ packages/babel-plugin-transform-runtime/test/fixtures/use-options/useESModules-cjs/output.js
@@ -1,0 +1,23 @@
+var _classCallCheck = require("@babel/runtime/helpers/classCallCheck");
+
+var _possibleConstructorReturn = require("@babel/runtime/helpers/possibleConstructorReturn");
+
+var _getPrototypeOf = require("@babel/runtime/helpers/getPrototypeOf");
+
+var _inherits = require("@babel/runtime/helpers/inherits");
+
+let Foo =
+/*#__PURE__*/
+function (_Bar) {
+  "use strict";
+
+  _inherits(Foo, _Bar);
+
+  function Foo() {
+    _classCallCheck(this, Foo);
+
+    return _possibleConstructorReturn(this, _getPrototypeOf(Foo).apply(this, arguments));
+  }
+
+  return Foo;
+}(Bar);
