+++ packages/babel-plugin-transform-runtime/test/fixtures/use-options/useESModules-mjs/input.mjs
@@ -1,0 +1,1 @@
+class Foo extends Bar {}
