+++ packages/babel-plugin-transform-runtime/test/fixtures/use-options/useESModules-mjs/options.json
@@ -1,0 +1,3 @@
+{
+  "plugins": [["transform-runtime", { "useESModules": true }], "transform-classes"]
+}
